backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
PyPDF2
Pillow
pydantic
aiofiles
huggingface-hub
transformers
torch
scipy
sentence-transformers
numpy
chromadb
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-docx
# Optional: faiss-cpu (fast HNSW indexing for Chroma)
